+++ src/node/create-webpack-config-client.js
@@ -4,10 +4,8 @@
 const _ = require('lodash');
 const path = require('path');
 const webpack = require('webpack');
-const crypto = require('crypto');
 const webpackMerge = require('webpack-merge');
 const AssetsPlugin = require('assets-webpack-plugin');
-const UglifyPlugin = require('uglifyjs-webpack-plugin');
 const resolveFrom = require('resolve-from');
 const createWebpackConfigBase = require('./create-webpack-config-base');
 const constants = require('./constants');
@@ -66,74 +64,15 @@ function createWebpackConfigClient(
         filename: 'assets.json',
         processOutput: (x) => JSON.stringify(x, null, 2)
       }),
-      // Extract universal vendor files (defined above) from everything else.
-      new webpack.optimize.CommonsChunkPlugin({
-        name: 'vendor',
-        minChunks: Infinity
-      }),
-      // Bundle together any other modules from anywhere imported more than 3 times.
-      new webpack.optimize.CommonsChunkPlugin({
-        name: 'app',
-        children: true,
-        minChunks: 4
-      }),
-      // Trying to follow advice for long-term caching described here:
-      // https://webpack.js.org/guides/caching/#extracting-boilerplate and
-      // https://jeremygayed.com/dynamic-vendor-bundling-in-webpack-528993e48aab#.hjgai17ap
-      // Because 'manifest' does not correspond to an entry name, this chunk
-      // will include Webpack's runtime boilerplate and manifest, which can
-      // change with each build. During the static build we inject it directly
-      // into the HTML, so those variations do not ruin caching on large chunks.
-      new webpack.optimize.CommonsChunkPlugin('manifest'),
-      // Recommended at https://webpack.js.org/guides/caching/#module-identifiers
-      // as a way to make module IDs more deterministic.
-      new webpack.HashedModuleIdsPlugin(),
-      new webpack.NamedChunksPlugin((chunk) => {
-        if (chunk.name) return chunk.name;
-        const hashSeed = Array.from(
-          chunk.modulesIterable,
-          (m) => m.resource || ''
-        )
-          .sort()
-          .map((resource) => {
-            return path.relative(batfishConfig.pagesDirectory, resource);
-          })
-          .join('_');
-        const chunkHash = crypto
-          .createHash('md5')
-          .update(hashSeed)
-          .digest('hex');
-        return chunkHash;
-      }), // Define an environment variable for special cases
       new webpack.DefinePlugin({
         'process.env.DEV_SERVER': (options && options.devServer) || false
       })
     ].concat(batfishConfig.webpackPlugins || []);
 
-    if (batfishConfig.production) {
-      const uglifyPlugin = new UglifyPlugin({
-        sourceMap: true,
-        cache: true,
-        parallel: true,
-        uglifyOptions: {
-          compress: {
-            // Disabled because of an issue with Uglify breaking seemingly valid code:
-            // https://github.com/facebookincubator/create-react-app/issues/2376
-            // Pending further investigation:
-            // https://github.com/mishoo/UglifyJS2/issues/2011
-            comparisons: false
-          },
-          output: {
-            // Turned on because emoji and regex is not minified properly using default
-            // https://github.com/facebookincubator/create-react-app/issues/2488
-            ascii_only: true
-          }
-        }
-      });
-      clientPlugins.push(uglifyPlugin);
-    }
-
-    const appEntry = [];
+    const appEntry = [
+      require.resolve('core-js/modules/es6.promise'),
+      require.resolve('core-js/modules/es6.array.iterator')
+    ];
     if (!batfishConfig.production && batfishConfig.inlineJs) {
       batfishConfig.inlineJs.forEach((jsData) => {
         appEntry.push(jsData.filename);
@@ -177,6 +116,22 @@ function createWebpackConfigClient(
         net: 'empty',
         tls: 'empty',
         child_process: 'empty'
+      },
+      optimization: {
+        moduleIds: 'hashed',
+        runtimeChunk: {
+          name: 'manifest'
+        },
+        splitChunks: {
+          cacheGroups: {
+            vendor: {
+              chunks: 'initial',
+              name: 'vendor',
+              test: 'vendor',
+              enforce: true
+            }
+          }
+        }
       }
     };